requirements-test.txt
# Test requirements for the PostgreSQL GCS Synchronizer
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
coverage>=7.0.0
mock>=4.0.0
psycopg2-binary>=2.9.0
google-cloud-storage>=2.10.0
python-dotenv>=1.0.0
